refactor: separate install function
