+++ packages/plugin-commands-installation/test/update/update.ts
@@ -1,6 +1,8 @@
+import PnpmError from '@pnpm/error'
 import { install, update } from '@pnpm/plugin-commands-installation'
-import prepare from '@pnpm/prepare'
+import prepare, { preparePackages } from '@pnpm/prepare'
 import { addDistTag } from '@pnpm/registry-mock'
+import path = require('path')
 import test = require('tape')
 import { DEFAULT_OPTS } from '../utils'
 
@@ -35,5 +37,36 @@ test('update with "*" pattern', async (t) => {
   t.ok(lockfile.packages['/peer-a/1.0.1'])
   t.ok(lockfile.packages['/peer-c/2.0.0'])
   t.ok(lockfile.packages['/pnpm-foo/1.0.0'])
+  t.end()
+})
+
+test('update: fail when both "latest" and "workspace" are true', async (t) => {
+  const projects = preparePackages(t, [
+    {
+      name: 'project-1',
+      version: '1.0.0',
+    },
+    {
+      name: 'project-2',
+      version: '2.0.0',
+    },
+  ])
+
+  let err!: PnpmError
+  try {
+    await update.handler(['project-2'], {
+      ...DEFAULT_OPTS,
+      dir: path.resolve('project-1'),
+      latest: true,
+      linkWorkspacePackages: false,
+      saveWorkspaceProtocol: false,
+      workspace: true,
+      workspaceDir: process.cwd(),
+    })
+  } catch (_err) {
+    err = _err
+  }
+  t.equal(err.code, 'ERR_PNPM_BAD_OPTIONS')
+  t.equal(err.message, 'Cannot use --latest with --workspace simultaneously')
   t.end()
 })
